Add hallucination score

+++ src/common/result.py
@@ -3,6 +3,6 @@ from common.passage import Passage
 
 
 class Result:
-    def __init__(self, query: str, passages: List[Tuple[Passage, int]]) -> None:
+    def __init__(self, query: str, passages: List[Tuple[Passage, float]]) -> None:
         self.query = query
         self.passages = passages
